Rename folder ometiff to microscopy
