+++ bids-validator/validators/bids/fullTest.js
@@ -6,7 +6,7 @@ import json from '../json'
 import NIFTI from '../nifti'
 import bval from '../bval'
 import bvec from '../bvec'
-import ometiff from '../ometiff'
+import ometiff from '../microscopy'
 import Events from '../events'
 import { session } from '../session'
 import checkAnyDataPresent from '../checkAnyDataPresent'
