+++ bids-validator/validators/microscopy/checkJSONAndField.js
