+++ bids-validator/validators/microscopy/index.js
