+++ bids-validator/validators/microscopy/ometiff.js
