+++ bids-validator/validators/microscopy/validate.js
